requirements.txt
requests
PyQt5
PyQt-Fluent-Widgets
tencentcloud-sdk-python